agent/requirements.txt
uagents>=0.12.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
uagents-core>=0.3.0
hyperon >= 0.2.8